refactor: change displayIndex type to long

+++ photo-service/src/main/java/kr/mafoo/photo/controller/dto/request/AlbumUpdateDisplayIndexRequest.java
@@ -9,6 +9,6 @@ public record AlbumUpdateDisplayIndexRequest(
         @NotNull
         @Min(0)
         @Schema(description = "대상 인덱스", example = "1")
-        Long newDisplayIndex
+        Integer newDisplayIndex
 ) {
 }
